make passing next to the publish optional
Relying on the publish function to call next for you is a little
dangerous and could cause fragility in the API if the publish borks out
for some reason.  Better to fire and forget and let the caller call next
so they are responsible for the ensuring their app moves to the next
step.